pyproject.toml
[tool.poetry]
name = "GPT-vision-fashionMnist"
version = "0.1.0"
description = ""
authors = ["Kirill Yakunin <yakunin.k@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
opencv-python = "^4.10.0.84"
urllib3 = "^2.2.2"
pillow = "^10.4.0"
openai = "^1.43.0"
tenacity = "^9.0.0"
python-dotenv = "^1.0.1"
torchvision = "^0.19.1"
scikit-learn = "^1.5.1"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
